requirements.txt
flask
python-dotenv
sentence-transformers
faiss-cpu
numpy
PyPDF2
google-genai